livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[mcp,silero,turn-detector]~=1.2",
    "livekit-plugins-deepgram>=1.2.14",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
    "tensorflow>=2.20.0",
    "httpx>=0.25.0",
    "supabase>=2.0.0",
    "openai>=1.0.0",
]
